pyproject.toml
[project]
name = "osx-echo"
version = "0.1.0"
description = "OSX Echo is a voice assistant that uses whisper.cpp for dictation."
authors = [
    { name = "Martin Vejmelka", email = "vejmelkam@gmail.com" }
]
dependencies = [
    "pyaudio>=0.2.14",
    "pynput>=1.7.7",
    "rumps>=0.4.0",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/osx_echo"]
